generate_pregnancy_flags

This function generates a table indicating whether each report meets the pregnancy-related criteria defined by the UMC, EMA, and Sakai algorithms for a given database.

Usage

generate_pregnancy_flags(database = FAERS_version, pids_vct = NA)

Arguments

database

A character string specifying the database to use. Default is FAERS_version. If "sample", sample data from the DiAna package is used.

pids_vct

A vector of primary IDs to filter the data. Default to NA, that takes all the primary IDs in the Demo table.

Value

A data.table containing the primary IDs, individual criteria flags, and aggregated flags for each algorithm. The UMC column indicates reports identified by the UMC/Vigibase algorithm; the EMA column by the EMA/EudraVigilance algorithm; and the SakaiHPPV, SakaiHS, and SakaiM columns by the Sakai algorithm's high PPV, high sensitivity, and medium versions, respectively.

Details

The function imports data from the specified database and checks various exclusion and inclusion criteria related to pregnancy. The criteria include age, sex, exposure, reactions, and outcomes.

Examples

if (FALSE) { # \dontrun{
# Example usage with sample data
pregnancy_retrieval(database = "sample", pids_vct = c(1, 2, 3))
} # }